big_vision/requirements.txt
numpy==1.23.5 
absl-py
huggingface-hub
flaxmodels
h5py
git+https://github.com/google/CommonLoopUtils
distrax
editdistance
einops
flax
optax
git+https://github.com/google/flaxformer
git+https://github.com/akolesnikoff/panopticapi.git@mute
overrides
protobuf==3.20.3
sentencepiece
tensorflow-cpu==2.12.0
tfds-nightly
tensorflow-text
tensorflow-gan
psutil
pycocoevalcap
webdataset>=0.2.48